Updated detector prefab
I had originally added a rigidbody to the detector prefab, but
things got really weird if I dropped that object inside another
one that has a rigidbody already.  In the case of the kittens,
the detector started wandering around really far from the model
when animating, with no apparent logic.

Currently it is the responsability of the user to ensure that
the parent object using the detector fits the requirements necessary
in order to receive trigger events.